+++ src/pages/NoticePage/CreateNoticePage/index.tsx
@@ -1,68 +1,21 @@
-import { useRef, useState, useEffect } from 'react';
-import { useNavigate } from 'react-router-dom';
-import { Editor } from '@toast-ui/react-editor';
 import RightContainer from '../../../components/layout/RightContainer';
 import TextEditor from '../../../components/features/TextEditor';
 import { TitleInputSection, ButtonContainer } from './style';
 import InputUI from '../../../components/ui/InputUI';
 import CheckboxUI from '../../../components/ui/CheckboxUI';
-import toast from 'react-hot-toast';
 import { DefaultButton } from '../../../components/ui/ButtonUI';
-import { FileUploaderUI } from '../../../components/ui/FileUploaderUI';
-import { useCreateNoticeRequest } from '../../../api-hooks/Notice';
-
-type EditorInstance = Editor | null;
+import useCreateNotice from './useCreateNotice';
 
 function CreateNoticePage() {
-  const navigate = useNavigate();
-  const editorRef = useRef<EditorInstance>(null);
-
-  const [title, setTitle] = useState('');
-  const [content, setContent] = useState('');
-  const [isTopChecked, setIsTopChecked] = useState(false);
-
-  const handleTitleInput = (e: React.ChangeEvent<HTMLInputElement>) => {
-    setTitle(e.target.value);
-  };
-
-  const handleContentInput = () => {
-    // eslint-disable-next-line @typescript-eslint/no-unsafe-argument, @typescript-eslint/no-unsafe-member-access, @typescript-eslint/no-unsafe-call
-    setContent(editorRef.current?.getInstance().getHTML());
-  };
-
-  const clickSelectTopNotice = () => {
-    setIsTopChecked((prevState) => !prevState);
-  };
-
   const {
-    mutate: adminCreateNoticeRequest,
-    data: adminCreateNoticeData,
-    error: adminCreateNoticeError,
-    isLoading: adminCreateNoticeLoading,
-  } = useCreateNoticeRequest();
-
-  const clickCreateNoticeSubmit = () => {
-    const data = {
-      title: title,
-      content: content,
-      top: isTopChecked,
-    };
-
-    if (!title || !content) {
-      toast.error('제목과 내용은 필수 항목 입니다.');
-    } else {
-      adminCreateNoticeRequest({ ...data });
-    }
-  };
-
-  useEffect(() => {
-    if (adminCreateNoticeData) {
-      toast.success('게시글 등록이 완료되었습니다.');
-      navigate('/notice/manage');
-    } else if (adminCreateNoticeError) {
-      toast.error((adminCreateNoticeError as Error).message);
-    }
-  }, [adminCreateNoticeData, adminCreateNoticeError]);
+    editorRef,
+    isTopChecked,
+    handleTitleInput,
+    handleContentInput,
+    clickSelectTopNotice,
+    clickCreateNoticeSubmit,
+    adminCreateNoticeLoading,
+  } = useCreateNotice();
 
   return (
     <RightContainer title={'공지사항 글 업로드'}>
