+++ src/pages/NoticePage/NoticeManagePage/index.tsx
@@ -1,4 +1,3 @@
-import { useState, useEffect } from 'react';
 import RightContainer from '../../../components/layout/RightContainer';
 import {
   NoticeSection,
@@ -16,71 +15,26 @@ import TextViewer from '../../../components/features/TextViewer';
 import Pagination from '../../../components/features/Pagination';
 import { SimpleGrid } from '@chakra-ui/react';
 import { RepeatIcon } from '@chakra-ui/icons';
-import { useGetNoticeRequest } from '../../../api-hooks/Notice';
-import toast from 'react-hot-toast';
-import { useSelectedArticleAtom } from '../../../store/articleAtom';
-import { changeDateFormat } from '../../../utils/changeDateFormat';
 import { DefaultButton } from '../../../components/ui/ButtonUI';
-import { INoticeData } from '../../../@types/Notice';
+import useNoticeManage from './useNoticeManage';
 
 function NoticeManagePage() {
-  const [noticeList, setNoticeList] = useState<INoticeData[] | null>([]);
-  const { selectedArticle, setSelectedArticle } = useSelectedArticleAtom();
-  const [currentPage, setCurrentPage] = useState(1);
-  const [totalPosts, setTotalPosts] = useState(1);
-  const [dateRange, setDateRange] = useState<[Date | null, Date | null]>([null, null]);
-  const [startDate, endDate] = dateRange;
-  const [keyword, setKeyword] = useState('');
-
-  const handleSearchKeyword = (e: React.ChangeEvent<HTMLInputElement>) => {
-    setKeyword(e.target.value);
-  };
-
   const {
-    data: noticeResult,
-    isError: noticeError,
-    refetch: noticeRefetch,
-  } = useGetNoticeRequest(
-    {
-      page: currentPage - 1,
-      startDate: startDate ? changeDateFormat(startDate, 'YYYY-MM-DDT00:00:00') : '',
-      endDate: endDate ? changeDateFormat(endDate, 'YYYY-MM-DDT23:59:59') : '',
-      title: keyword ? encodeURIComponent(keyword) : '',
-    },
-    false,
-  );
-
-  useEffect(() => {
-    void noticeRefetch();
-  }, [currentPage]);
-
-  const handleDateSearchButton = () => {
-    void noticeRefetch();
-  };
-
-  const handleOnEnterKeyDown = (e: React.KeyboardEvent<HTMLInputElement>) => {
-    if (e.key === 'Enter') {
-      void noticeRefetch();
-    }
-  };
-
-  useEffect(() => {
-    if (noticeResult) {
-      setNoticeList(noticeResult?.data.noticeList);
-      setTotalPosts(noticeResult?.data.totalPage);
-    } else if (noticeError) {
-      toast.error('정보 글 목록을 불러오는데 실패했습니다.');
-    }
-  }, [noticeResult, noticeError]);
-
-  const handleNoticePreview = (notice: INoticeData) => {
-    setSelectedArticle(notice);
-  };
-
-  const handleSearchRefreshButton = () => {
-    setDateRange([null, null]);
-    setKeyword('');
-  };
+    selectedArticle,
+    currentPage,
+    setCurrentPage,
+    setDateRange,
+    startDate,
+    endDate,
+    keyword,
+    handleSearchKeyword,
+    noticeList,
+    totalPosts,
+    handleDateSearchButton,
+    handleOnEnterKeyDown,
+    handleNoticePreview,
+    handleSearchRefreshButton,
+  } = useNoticeManage();
 
   return (
     <RightContainer title={'공지사항 글 관리'}>
